Refine mypy type hint in _set_fine_search_range function
- Update type hint for `min_entropy_ind` from built-in `int` to `np.intp`.
  - This refinement addresses a type compatibility warning.
  - `np.intp` is the integer type used by numpy for indexing and can differ in size from the built-in Python `int` depending on the platform (32-bit vs 64-bit).
  - Mypy highlighted this type hint discrepancy.

+++ cleanlab/internal/multiannotator_utils.py
@@ -323,7 +323,7 @@ def find_best_temp_scaler(
 
 
 def _set_fine_search_range(
-    coarse_search_range: list, fine_search_size: int, min_entropy_ind: int
+    coarse_search_range: list, fine_search_size: int, min_entropy_ind: np.intp
 ) -> np.ndarray:
     fine_search_range = np.array([])
     if min_entropy_ind != 0:
